LWMD-16 | New search service methods
