+++ Page_Library/Search/Service/Interface/IPageSearchService.cs
@@ -1,16 +1,13 @@
-﻿using Page_Library.Search.Entities.SearchResult.Interface;
+﻿using Page_Library.Content.Repository.Interface;
+using Page_Library.Page.Entities.Page.Interface;
+using Page_Library.Page.Repository.Interface;
+using Page_Library.Search.Entities.SearchResult.Interface;
 
 namespace Page_Library.Search.Service.Interface
 {
     public interface IPageSearchService
     {
 
-        public abstract List<ISearchResult> Search(string searchTerm,
-            bool programming,
-            bool testing,
-            bool games,
-            bool threeDAssets,
-            bool twoDAssets,
-            bool blog);
+        public List<ISearchResult> Search(string searchTerm, string category);
     }
 }
